% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.amber}
\alias{bv.amber}
\title{Brand Victoria amber}
\format{
## `bv.amber`
An object of class `colour` with hex value `#0072ce`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.amber
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The third colour in the seventh secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.amber,3))
}
\keyword{datasets}
