% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.azure}
\alias{bv.azure}
\title{Brand Victoria azure}
\format{
## `bv.azure`
An object of class `colour` with hex value `#00a9e0`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.azure
}
\description{
The second colour in the second secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.azure,3))
}
\keyword{datasets}
