% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.black}
\alias{bv.black}
\title{Brand Victoria black}
\format{
## `bv.black`
An object of class `colour` with hex value `#000000`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.black
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.black,3))
}
\keyword{datasets}
