% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.blush}
\alias{bv.blush}
\title{Brand Victoria blush}
\format{
## `bv.blush`
An object of class `colour` with hex value `#ef4a81`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.blush
}
\description{
The second colour in the tenth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.blush,3))
}
\keyword{datasets}
