% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.butter}
\alias{bv.butter}
\title{Brand Victoria butter}
\format{
## `bv.butter`
An object of class `colour` with hex value `#fdda24`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.butter
}
\description{
The second colour in the sixth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.butter,3))
}
\keyword{datasets}
