% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.canary}
\alias{bv.canary}
\title{Brand Victoria canary}
\format{
## `bv.canary`
An object of class `colour` with hex value `#f3e500`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.canary
}
\description{
The secondc olour in the fifth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.canary,3))
}
\keyword{datasets}
