% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.cerulean}
\alias{bv.cerulean}
\title{Brand Victoria cerulean}
\format{
## `bv.cerulean`
An object of class `colour` with hex value `#0090da`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.cerulean
}
\description{
The first colour in the second secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.cerulean,3))
}
\keyword{datasets}
