% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.charcoal}
\alias{bv.charcoal}
\title{Brand Victoria charcoal}
\format{
## `bv.charcoal`
An object of class `colour` with hex value `#53565a`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.charcoal
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.charcoal,3))
}
\keyword{datasets}
