% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.chartreuse}
\alias{bv.chartreuse}
\title{Brand Victoria chartreuse}
\format{
## `bv.chartreuse`
An object of class `colour` with hex value `#cedc00`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.chartreuse
}
\description{
The secondc olour in the fifth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.chartreuse,3))
}
\keyword{datasets}
