% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.cobalt}
\alias{bv.cobalt}
\title{Brand Victoria cobalt}
\format{
## `bv.cobalt`
An object of class `colour` with hex value `#0072CE`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.cobalt
}
\description{
The third colour in the first secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.cobalt,3))
}
\keyword{datasets}
