% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.crystal}
\alias{bv.crystal}
\title{Brand Victoria crystal}
\format{
## `bv.crystal`
An object of class `colour` with hex value `#88dbdf`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.crystal
}
\description{
The third colour in the third secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.crystal,3))
}
\keyword{datasets}
