% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.cyan}
\alias{bv.cyan}
\title{Brand Victoria cyan}
\format{
## `bv.cyan`
An object of class `colour` with hex value `#009ca6`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.cyan
}
\description{
The first colour in the third secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.cyan,3))
}
\keyword{datasets}
