% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.fire}
\alias{bv.fire}
\title{Brand Victoria fire}
\format{
## `bv.fire`
An object of class `colour` with hex value `#E57200`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.fire
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The third colour in the seventh secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.fire,3))
}
\keyword{datasets}
