% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.forest}
\alias{bv.forest}
\title{Brand Victoria forest}
\format{
## `bv.forest`
An object of class `colour` with hex value `#00573F`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.forest
}
\description{
The first colour in the fourth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.forest,3))
}
\keyword{datasets}
