% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.jade}
\alias{bv.jade}
\title{Brand Victoria jade}
\format{
## `bv.jade`
An object of class `colour` with hex value `#007b4b`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.jade
}
\description{
The second colour in the fourth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.jade,3))
}
\keyword{datasets}
