% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.maroon}
\alias{bv.maroon}
\title{Brand Victoria maroon}
\format{
## `bv.maroon`
An object of class `colour` with hex value `#8A2A2B`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.maroon
}
\description{
The first colour in the eigth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.maroon,3))
}
\keyword{datasets}
