% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.mint}
\alias{bv.mint}
\title{Brand Victoria mint}
\format{
## `bv.mint`
An object of class `colour` with hex value `#00B140`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.mint
}
\description{
The third colour in the fourth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.mint,3))
}
\keyword{datasets}
