% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.musk}
\alias{bv.musk}
\title{Brand Victoria musk}
\format{
## `bv.musk`
An object of class `colour` with hex value `#e89cae`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.musk
}
\description{
The third colour in the 10th secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.musk,3))
}
\keyword{datasets}
