% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.navy}
\alias{bv.navy}
\title{Brand Victoria navy}
\format{
## `bv.navy`
An object of class `colour` with hex value `#201547`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.navy
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The first colour in the 1st secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.navy,3))
}
\keyword{datasets}
