% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.orange}
\alias{bv.orange}
\title{Brand Victoria orange}
\format{
## `bv.orange`
An object of class `colour` with hex value `#E35205`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.orange
}
\description{
The first colour in the seventh secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.orange,3))
}
\keyword{datasets}
