% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.pink}
\alias{bv.pink}
\title{Brand Victoria pink}
\format{
## `bv.pink`
An object of class `colour` with hex value `#c63663`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.pink
}
\description{
The first colour in the 10th secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.pink,3))
}
\keyword{datasets}
