% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.plum}
\alias{bv.plum}
\title{Brand Victoria plum}
\format{
## `bv.plum`
An object of class `colour` with hex value `#642667`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.plum
}
\description{
The second colour in the ninth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.plum,3))
}
\keyword{datasets}
