% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.purple}
\alias{bv.purple}
\title{Brand Victoria purple}
\format{
## `bv.purple`
An object of class `colour` with hex value `#87189D`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.purple
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The third colour in the ninth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.purple,3))
}
\keyword{datasets}
