% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.red}
\alias{bv.red}
\title{Brand Victoria red}
\format{
## `bv.red`
An object of class `colour` with hex value `#AF272F`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.red
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The second colour in the eigth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.red,3))
}
\keyword{datasets}
