% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.rose}
\alias{bv.rose}
\title{Brand Victoria rose}
\format{
## `bv.rose`
An object of class `colour` with hex value `#D50032`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.rose
}
\description{
The third colour in the eigth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.rose,3))
}
\keyword{datasets}
