% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.royal}
\alias{bv.royal}
\title{Brand Victoria royal}
\format{
## `bv.royal`
An object of class `colour` with hex value `#004c97`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.royal
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The second colour in the first secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.royal,3))
}
\keyword{datasets}
