% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.sky}
\alias{bv.sky}
\title{Brand Victoria sky}
\format{
## `bv.sky`
An object of class `colour` with hex value `#0090da`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.sky
}
\description{
The third colour in the second secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.sky,3))
}
\keyword{datasets}
