% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.smoke}
\alias{bv.smoke}
\title{Brand Victoria smoke}
\format{
## `bv.smoke`
An object of class `colour` with hex value `#d9d9d6`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.smoke
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.smoke,3))
}
\keyword{datasets}
