% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.sunshine}
\alias{bv.sunshine}
\title{Brand Victoria sunshine}
\format{
## `bv.sunshine`
An object of class `colour` with hex value `#F6BE00`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.sunshine
}
\description{
The first colour in the fifth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.sunshine,3))
}
\keyword{datasets}
