% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.teal}
\alias{bv.teal}
\title{Brand Victoria teal}
\format{
## `bv.teal`
An object of class `colour` with hex value `#00b2a9`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.teal
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\details{
The second colour in the third secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.teal,3))
}
\keyword{datasets}
