% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.violet}
\alias{bv.violet}
\title{Brand Victoria violet}
\format{
## `bv.violet`
An object of class `colour` with hex value `#500778`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.violet
}
\description{
The first colour in the ninth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.violet,3))
}
\keyword{datasets}
