% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.white}
\alias{bv.white}
\title{Brand Victoria white}
\format{
## `bv.white`
An object of class `colour` with hex value `#FFFFFF`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.white
}
\description{
One of ten primary colours in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.white,3))
}
\keyword{datasets}
