% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.yellow}
\alias{bv.yellow}
\title{Brand Victoria yellow}
\format{
## `bv.yellow`
An object of class `colour` with hex value `#fbe122`.
}
\source{
<https://www.vic.gov.au/brand-victoria-guidelines>
}
\usage{
bv.yellow
}
\description{
The third colour in the sixth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.yellow,3))
}
\keyword{datasets}
