% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv_colours}
\alias{bv_colours}
\title{Brand Victoria colours}
\format{
## `bv_colours`
An object of class `palette` with `r length(bv_colours)` palettes.
Use `names(bv_colours)` to return all palette names.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv_colours
}
\description{
Brand Victoria colour palettes.
}
\details{
The Brand Victoria colour palette consists of ten primary colours
and ten secondary colour groups that incorporate some primary
colours.

`bv_colours` includes these as named palettes as well as a palette of
all Brand Victoria colours defined in this package.
}
\examples{
# Plot all colours
barplot(c(2,5,9,7,4,3,1,8,6), col=bv_colours$primary)
}
\keyword{datasets}
