% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_department_colours.R
\docType{data}
\name{dffh_colours}
\alias{dffh_colours}
\title{Department of Families, Fairness and House colours}
\format{
## `dffh_colours`
An object of class `palette` with `r length(dffh_colours)` palettes and
`r length(unlist(dffh_colours))-4` colours.

Use `names(dffh_colours)` to return all palette names.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
dffh_colours
}
\description{
Department of Families, Fairness and House colour palettes.
}
\details{
The departments colour palette is based on the Brand Victoria colours,
utilising:

- `bv.teal`, `bv.mint`, `bv.cerulean`, and `bv.purple` as `core` colours.
- `bv.navy`, `bv.charcoal`, and `bv.smoke` as `support` colours.
- `bv.orange`, `bv.pink` and `bv.red` as `extended` colours.

A range of tints based on the `core` colours are also included in the palette
e.g. `tints_teal`
}
\examples{
# Plot primary colours
barplot(c(2,5,9,7), col=dffh_colours$core)
# Plot pink tints
barplot(c(2,5,9,7), col=dffh_colours$tints_teal)
}
\keyword{datasets}
