% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_department_colours.R
\docType{data}
\name{dh_colours}
\alias{dh_colours}
\title{Department of Health colours}
\format{
## `dh_colours`
An object of class `palette` with `r length(dh_colours)` palettes.
Use `names(dh_colours)` to return all palette names.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
dh_colours
}
\description{
Department of Health colour palettes.
}
\details{
The departments colour palette is based on the Brand Victoria colours,
utilising:

- `bv.pink`, `bv.maroon`, `bv.orange`, and `bv.royal` as `primary` colours.
- `bv.blush`, `bv.rose`, `bv.amber`, and `bv.cobalt` as `secondary` colours.

A range of tints based on the primary colours are also included in the palette
e.g. `tints_pink`
}
\examples{
# Plot primary colours
barplot(c(2,5,9,7), col=dh_colours$primary)
# Plot pink tints
barplot(c(2,5,9,7), col=dh_colours$tints_pink)
}
\keyword{datasets}
