% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_department_colours.R
\docType{data}
\name{djcs_colours}
\alias{djcs_colours}
\title{Department of Justice and Community Safety colours}
\format{
## `djcs_colours`
An object of class `palette` with `r length(djcs_colours)` palettes.
Use `names(djcs_colours)` to return all palette names.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
djcs_colours
}
\description{
Department of Justice and Community Safety colour palettes.
}
\details{
The departments colour palette is based on the Brand Victoria colours,
utilising:

- `bv.navy`, `bv.sky`, `bv.violet`, `bv.cobalt`, `bv.charcoal`, `bv.smoke` as `primary` colours.
A range of tints based on the primary colours are also included in the palette
e.g. `tints_navy`
}
\examples{
# Plot primary colours
barplot(c(2,5,9,7), col=djcs_colours$primary)
# Plot pink tints
barplot(c(2,5,9,7), col=djcs_colours$tints_purple)
}
\keyword{datasets}
