% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_vic_fonts.R
\name{install_rubik_fonts}
\alias{install_rubik_fonts}
\title{Install Rubik Fonts}
\usage{
install_rubik_fonts()
}
\description{
Themes in this package use the Rubik font as an alternative to VIC.
}
\details{
This function allow this font to be installed within the package
}
\seealso{
Other vpstheme: 
\code{\link{coord_vps_map}()},
\code{\link{install_vic_fonts}()},
\code{\link{make_tints}()},
\code{\link{scale_colour_dffh}()},
\code{\link{scale_colour_dffh_line}()},
\code{\link{scale_colour_dh}()},
\code{\link{scale_colour_dh_line}()},
\code{\link{scale_colour_djcs}()},
\code{\link{scale_colour_djcs_line}()},
\code{\link{scale_fill_dffh}()},
\code{\link{scale_fill_dh}()},
\code{\link{scale_fill_djcs}()},
\code{\link{theme_vps_dffh}()},
\code{\link{theme_vps_dh}()},
\code{\link{theme_vps_djcs}()},
\code{\link{theme_vps_foundation}()}
}
\concept{vpstheme}
