% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_vic_fonts.R
\name{install_vic_fonts}
\alias{install_vic_fonts}
\title{Install VIC Fonts}
\usage{
install_vic_fonts()
}
\description{
Themes in this package use the VIC font.
}
\details{
This function when complete will allow these to be installed
within the package
}
\seealso{
Other vpstheme: 
\code{\link{install_rubik_fonts}()},
\code{\link{make_tints}()},
\code{\link{scale_colour_dffh}()},
\code{\link{scale_colour_dffh_line}()},
\code{\link{scale_colour_dh}()},
\code{\link{scale_colour_dh_line}()},
\code{\link{scale_fill_dffh}()},
\code{\link{scale_fill_dh}()},
\code{\link{theme_vps_dffh}()},
\code{\link{theme_vps_dh}()},
\code{\link{theme_vps_foundation}()}
}
\concept{vpstheme}
