% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_vic_lga.R
\name{lga_string_tidy}
\alias{lga_string_tidy}
\title{Helper function used by clean_lga_name}
\usage{
lga_string_tidy(string)
}
\arguments{
\item{string}{a character string.}
}
\description{
Helper function used by clean_lga_name
}
