% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_vps_dffh.R
\name{scale_colour_dffh_line}
\alias{scale_colour_dffh_line}
\title{Department of Families, Fairness and Housing color scale function for line charts}
\usage{
scale_colour_dffh_line(
  ...,
  colour = c("teal", "mint", "blue", "purple"),
  gradient = FALSE,
  lighten = FALSE
)
}
\arguments{
\item{...}{Arguments passed on to scale_colour_manual}

\item{colour}{A text string that is used to select the palette to use}

\item{gradient}{A logical value string that is used to select the palette to use}

\item{lighten}{A logical value string that is used to select the palette to use}
}
\description{
This function is a wrapper around \code{\link[ggplot2]{scale_colour_manual}()} that supplies a
primary colour  \code{\link[vpstheme]{dh_colours}()} to create a discrete colour scale comprising
that colour and four grey colours.
}
\details{
By default the greys are \code{\link[vpstheme]{bv.charcoal}()}.
Setting `lighten = TRUE` changes them to \code{\link[vpstheme]{bv.smoke}()}
Setting  `gradient = TRUE` creates a palette ramp from \code{\link[vpstheme]{bv.black}()}
using \code{\link[vpstheme]{make_tints}()}
}
\seealso{
Other vpstheme: 
\code{\link{coord_vps_map}()},
\code{\link{install_rubik_fonts}()},
\code{\link{install_vic_fonts}()},
\code{\link{make_tints}()},
\code{\link{scale_colour_dffh}()},
\code{\link{scale_colour_dh}()},
\code{\link{scale_colour_dh_line}()},
\code{\link{scale_fill_dffh}()},
\code{\link{scale_fill_dh}()},
\code{\link{theme_vps_dffh}()},
\code{\link{theme_vps_dh}()},
\code{\link{theme_vps_foundation}()}
}
\concept{vpstheme}
