% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vps_dffh.R
\name{theme_vps_dffh}
\alias{theme_vps_dffh}
\title{Department of Families Fairness and Housing theme}
\usage{
theme_vps_dffh(
  base_colour = c("white", "teal", "mint", "blue", "purple"),
  base_size = 15
)
}
\arguments{
\item{base_colour}{A text string that is used to set background, line and text colours for the theme}

\item{base_size}{An integer that is used to set the base font size}
}
\description{
This theme takes \code{\link[vpstheme]{theme_vps_foundation}()} and applies
background colour and text colours based on parameter `base_colour`.
}
\details{
Colours are sourced from `vpstheme` and are Department style guide compliant.
}
\seealso{
Other vpstheme: 
\code{\link{install_rubik_fonts}()},
\code{\link{install_vic_fonts}()},
\code{\link{make_tints}()},
\code{\link{scale_colour_dh}()},
\code{\link{scale_fill_dh}()},
\code{\link{theme_vps_dh}()},
\code{\link{theme_vps_foundation}()}
}
\concept{vpstheme}
