% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_brand_victoria.R
\docType{data}
\name{bv.lime}
\alias{bv.lime}
\title{Brand Victoria lime}
\format{
## `bv.lime`
An object of class `colour` with hex value `#78BE20`.
}
\source{
<https://github.com/zerogetsamgow/vpstheme>
}
\usage{
bv.lime
}
\description{
The first colour in the fifth secondary colour group
in the Brand Victoria colour palette.
}
\examples{
barplot(c(2,5,3), col=rep(bv.lime,3))
}
\keyword{datasets}
