% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_vps_dh.R
\name{theme_vps_dh}
\alias{theme_vps_dh}
\title{Department of Health Theme}
\usage{
theme_vps_dh(
  base_colour = c("white", "pink", "orange", "blue", "red"),
  base_size = 15,
  show_gridlines = FALSE
)
}
\arguments{
\item{base_colour}{A text string that is used to set background, line and text colours for the theme}

\item{base_size}{An integer that is used to set the base font size}

\item{show_gridlines}{Add gridlines on the y-axis, defaults to FALSE. Use TRUE for gridlines.}
}
\description{
This theme takes \code{\link[vpstheme]{theme_vps_foundation}()} and applies
background colour and text colours based on parameter `base_colour`.
}
\details{
Colours are sourced from `vpstheme` and are Department style guide compliant.
}
\seealso{
Other vpstheme: 
\code{\link{coord_vps_map}()},
\code{\link{install_rubik_fonts}()},
\code{\link{install_vic_fonts}()},
\code{\link{make_tints}()},
\code{\link{scale_colour_dffh}()},
\code{\link{scale_colour_dffh_line}()},
\code{\link{scale_colour_dh}()},
\code{\link{scale_colour_dh_line}()},
\code{\link{scale_fill_dffh}()},
\code{\link{scale_fill_dh}()},
\code{\link{theme_vps_dffh}()},
\code{\link{theme_vps_foundation}()}
}
\concept{vpstheme}
